Resolve some run-time warnings

+++ aom_dsp/avg.c
@@ -463,8 +463,6 @@ int16_t aom_int_pro_col_c(const uint8_t *ref, const int width) {
   return sum;
 }
 
-// ref: [0 - 510]
-// src: [0 - 510]
 // bwl: {2, 3, 4, 5}
 int aom_vector_var_c(const int16_t *ref, const int16_t *src, const int bwl) {
   int i;
@@ -472,12 +470,12 @@ int aom_vector_var_c(const int16_t *ref, const int16_t *src, const int bwl) {
   int sse = 0, mean = 0, var;
 
   for (i = 0; i < width; ++i) {
-    int diff = ref[i] - src[i];  // diff: dynamic range [-510, 510], 10 bits.
-    mean += diff;                // mean: dynamic range 16 bits.
-    sse += diff * diff;          // sse:  dynamic range 26 bits.
+    int diff = ref[i] - src[i];
+    mean += diff;
+    sse += diff * diff;
   }
 
-  // (mean * mean): dynamic range 31 bits.
-  var = sse - ((mean * mean) >> (bwl + 2));
+  // TODO(any): check (mean * mean) dynamic range in bits.
+  var = sse - (int)(((int64_t)mean * (int64_t)mean) >> (bwl + 2));
   return var;
 }
